+++ backend/package.json
@@ -23,6 +23,7 @@
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "express-basic-auth": "^1.2.0",
     "express-rate-limit": "^5.1.3",
     "helmet": "^4.2.0",
     "js-yaml": "^3.14.1",
